Move simulator tests to appropriate module
